feat: add global fire animations and update streak counter
