+++ src/app/globals.css
@@ -8,6 +8,38 @@
   .text-balance {
     text-wrap: balance;
   }
+  
+  @keyframes fire-glow {
+    0%, 100% {
+      box-shadow: 0 0 5px 2px rgba(255, 92, 92, 0.7);
+      border-color: rgba(255, 92, 92, 0.7);
+    }
+    50% {
+      box-shadow: 0 0 15px 5px rgba(255, 165, 0, 0.9);
+      border-color: rgba(255, 165, 0, 0.9);
+    }
+  }
+  
+  .animate-fire-glow {
+    animation: fire-glow 2s infinite alternate;
+    position: relative;
+  }
+  
+  @keyframes flicker {
+    0%, 100% {
+      filter: drop-shadow(0 0 4px rgba(255, 92, 92, 0.5));
+      transform: scale(1);
+    }
+    50% {
+      filter: drop-shadow(0 0 8px rgba(255, 165, 0, 0.7));
+      transform: scale(1.05);
+    }
+  }
+  
+  .fire-animation {
+    animation: flicker 1.5s infinite alternate;
+    transform-origin: center;
+  }
 }
 
 .loader {
